testing impl of casting sub-types with a single fixed field supported.
Bytes still are read as a raw byte[], when really they are base64 encoded. We need to map that incoming type to
the encoded version, which means creating a new output vector.

Other types should work just fine as casts.

We also need to support ignoring the Tn|| prefixed fields with known type as they are propagated with
ANY schema (often leading to VarChar vectors)