Using drill MapConverter tools to do vector transfer.
Based on the ScanBatch implementation. I think its a bit heavy copying over bytes for all the fields, when really we just want to transfer the bytes
for a small subset of the non-null columns.

We could be smarter at the very least for BASE columns

There is a bunch of cruft in this commit from alternative ideas that should be stripped, but is retained in case this was a bad idea.