Correctly implementing reading info schema/connection metadata
Changed implementation to not be a wrapper around the returned
Drill metadata and attempt to translate it, but rather
doing a deeper integration with Drill in the adapter
that supports actually hiding and limiting access
based on the username.

Now we also pass in the orgID as the username to ensure
that the query is always bound correctly. For now, just
used for metadata filtering. Customer still cannot
access other tables because we prefix with the orgID
for specific table access in the ReParser.

There is also a little bit of change in how queries are
executed for e2e testing - we no longer generate the SQL
statement from the org/metric, but instead rely on the
outside query provider to generate the correct statement.

Closes #16