BUG: Fixing reads where there is a missing column.
Not sure how this didn't come up before. Basically, if we write some data, add a new column
and then read all the data, we end up with a NPE for the column for the 'older' rows
because there is no vector in the map for that timestamp.

This fix adds a client-like test to cover this case and fixes the bug